Sarthak | Changes the signature of Now method in Clock interface and FileIdGenerator now uses clock to generate id

+++ clock/Clock.go
@@ -3,7 +3,7 @@ package clock
 import "time"
 
 type Clock interface {
-	Now() int
+	Now() int64
 }
 
 type SystemClock struct{}
@@ -12,6 +12,6 @@ func NewSystemClock() *SystemClock {
 	return &SystemClock{}
 }
 
-func (clock *SystemClock) Now() int {
-	return int(time.Now().UnixNano())
+func (clock *SystemClock) Now() int64 {
+	return time.Now().UnixNano()
 }
